// front/services/models/loop-step.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Global Testing
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import { Condition } from './condition';

/**
 * 
 * @export
 * @interface LoopStep
 */
export interface LoopStep {
    /**
     * 
     * @type {string}
     * @memberof LoopStep
     */
    'scenario': string;
    /**
     * 
     * @type {string}
     * @memberof LoopStep
     */
    'id'?: string;
    /**
     * 
     * @type {string}
     * @memberof LoopStep
     */
    'name': string;
    /**
     * 
     * @type {number}
     * @memberof LoopStep
     */
    'stepNumber': number;
    /**
     * 
     * @type {{ [key: string]: string; }}
     * @memberof LoopStep
     */
    'variables'?: { [key: string]: string; };
    /**
     * 
     * @type {boolean}
     * @memberof LoopStep
     */
    'runAfterFailure'?: boolean;
    /**
     * 
     * @type {Array<Condition>}
     * @memberof LoopStep
     */
    'conditions'?: Array<Condition>;
    /**
     * 
     * @type {number}
     * @memberof LoopStep
     */
    'stepToReturn'?: number;
}