// front/services/models/sleep-step.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Global Testing
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface SleepStep
 */
export interface SleepStep {
    /**
     * 
     * @type {string}
     * @memberof SleepStep
     */
    'scenario': string;
    /**
     * 
     * @type {string}
     * @memberof SleepStep
     */
    'id'?: string;
    /**
     * 
     * @type {string}
     * @memberof SleepStep
     */
    'name': string;
    /**
     * 
     * @type {number}
     * @memberof SleepStep
     */
    'stepNumber': number;
    /**
     * 
     * @type {{ [key: string]: string; }}
     * @memberof SleepStep
     */
    'variables'?: { [key: string]: string; };
    /**
     * 
     * @type {boolean}
     * @memberof SleepStep
     */
    'runAfterFailure'?: boolean;
    /**
     * 
     * @type {number}
     * @memberof SleepStep
     */
    'duration'?: number;
}